+++ src/transformers/bunny.ts
@@ -1,8 +1,13 @@
 import { UrlParser, UrlTransformer } from "../types.ts";
-import { getNumericParam, setParamIfDefined } from "../utils.ts";
+import {
+  getNumericParam,
+  setParamIfDefined,
+  setParamIfUndefined,
+  toUrl,
+} from "../utils.ts";
 
 export const parse: UrlParser<{ fit?: string }> = (url) => {
-  const parsedUrl = new URL(url);
+  const parsedUrl = toUrl(url);
 
   const width = getNumericParam(parsedUrl, "width");
   const height = getNumericParam(parsedUrl, "height");
@@ -23,8 +28,10 @@ export const parse: UrlParser<{ fit?: string }> = (url) => {
 export const transform: UrlTransformer = (
   { url: originalUrl, width, height },
 ) => {
-  const url = new URL(originalUrl);
+  const url = toUrl(originalUrl);
   setParamIfDefined(url, "width", width, true, true);
-  setParamIfDefined(url, "height", height, true, true);
+  if (width && height) {
+    setParamIfUndefined(url, "aspect_ratio", `${width}:${height}`);
+  }
   return url;
 };